# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "一款面向小型团队的任务进度共享软件",
        "title": "Team",
        "termsOfService": "http://swagger.io/terrms",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:9918",
    "paths": {
...
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "\"登录\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登陆成功"
                    },
                    "400": {
                        "description": "输入格式错误"
                    },
                    "401": {
                        "description": "密码错误"
                    },
                    "404": {
                        "description": "用户不存在"
                    }
                }
            }
        },
...
        "/user": {
            "post": {
                "description": "\"注册一个新用户\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "\"注册\"",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户创建成功"
                    },
                    "400": {
                        "description": "输入有误，格式错误"
                    },
                    "401": {
                        "description": "电话号码重复"
                    }
                }
            }
        },
        "/user/avatar": {
            "put": {
                "description": "\"修改用户头像\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "\"修改头像\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功"
                    },
                    "400": {
                        "description": "上传失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/user/change_password/change": {
            "post": {
                "description": "“修改密码”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userinfo"
                ],
                "summary": "“修改用户密码”",
                "parameters": [
                    {
                        "description": "输入两次新密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Password"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "修改失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/user/change_password/verify": {
            "get": {
                "description": "“修改密码前对密码的验证功能”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userinfo"
                ],
                "summary": "“验证用户密码”",
                "parameters": [
                    {
                        "description": "输入密码",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证成功"
                    },
                    "400": {
                        "description": "输入有误"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/user/feedback": {
            "put": {
                "description": "“用户反馈”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userinfo"
                ],
                "summary": "“用户反馈”",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "输入反馈信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "反馈成功"
                    },
                    "400": {
                        "description": "输入格式有误"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/user/info": {
            "put": {
                "description": "“修改用户的基本信息”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userinfo"
                ],
                "summary": "“修改用户的信息”",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "输入昵称",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "修改失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/user/pupup": {
            "post": {
                "description": "\"再点击“完成设置”之前头像已经设置完，注册后弹窗里输入昵称\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "\"初始化用户信息\"",
                "parameters": [
                    {
                        "description": "输入昵称",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功"
                    },
                    "400": {
                        "description": "输入有误"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/user/pupup/avatar": {
            "post": {
                "description": "\"修改用户头像\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "\"修改头像\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功"
                    },
                    "400": {
                        "description": "上传失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Password": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "feedback": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                }
            }
        },
...
    }
}